- re factor some widget.
