--- lib/futures/timer/presentation/functions/close.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-void close(BuildContext context) {
-  SystemChannels.platform.invokeMethod('SystemNavigator.pop');
-}
