+++ lib/shared/presentation/functions/close.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+void close(BuildContext context) {
+  context.pop();
+}
